% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRCluster.R
\name{defaultGraalVMHome}
\alias{defaultGraalVMHome}
\title{Default GraalVM installation path}
\usage{
defaultGraalVMHome()
}
\value{
The default GraalVM installation path
}
\description{
Gives the path to the default location of GraalVM installation that includes FastR.
The default location is inside the directory where the fastRCluster was installed.
}
\details{
\code{\link{getGraalVMHome}} uses this value as the default,
if no other value is explicitly configured via R options or an environment variable.
}
\seealso{
\code{\link{getGraalVMHome}}
}
