% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRCluster.R
\name{fastRClusterInstallPackages}
\alias{fastRClusterInstallPackages}
\title{Installs packages on the FastR engine}
\usage{
fastRClusterInstallPackages(...)
}
\arguments{
\item{...}{Parameters passed to the R function \code{install.packages} that is run on the FastR engine.}
}
\value{
Invisible \code{NULL}
}
\description{
Installs packages on the FastR engine
}
