% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRCluster.R
\name{getGraalVMHome}
\alias{getGraalVMHome}
\title{Currently configured GraalVM path}
\usage{
getGraalVMHome()
}
\value{
The currently configured path to GraalVM installation.
}
\description{
Gives the path that is used as a default value of the \code{graalVMHome} parameter
for most of the functions in the fastRCluster package.
}
\details{
The value is taken from (in this order)
\enumerate{
    \item R option "graalvm.home"
    \item environment variable \code{GRAALVM_HOME}
    \item \code{\link{defaultGraalVMHome}()}
}
}
\seealso{
\code{\link{defaultGraalVMHome}}
}
