% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRCluster.R
\name{installFastR}
\alias{installFastR}
\title{Installs FastR}
\usage{
installFastR(path = defaultGraalVMHome())
}
\arguments{
\item{path}{Path to a directory where GraalVM should be installed. Defaults to \code{\link{defaultGraalVMHome}()}.}
}
\value{
the path where GraalVM was installed if successful (invisible), otherwise this function raises an error.
}
\description{
Downloads GraalVM Community Edition and installs the R ("FastR") component for GraalVM.
}
\details{
Note: the download is around 300MB. The installation usually takes few seconds.
If the given directory already contains GraalVM, this function installs the R ("FastR") component.
}
\seealso{
\code{\link{defaultGraalVMHome}}
}
